+++ scalyr_agent/copying_manager/worker.py
@@ -507,8 +507,10 @@ class CopyingManagerWorkerSession(
 
                         # Collect log lines to send if we don't have one already.
                         if self.__pending_add_events_task is None:
-                            self.__pending_add_events_task = self.__get_next_add_events_task(
-                                copying_params.current_bytes_allowed_to_send
+                            self.__pending_add_events_task = (
+                                self.__get_next_add_events_task(
+                                    copying_params.current_bytes_allowed_to_send
+                                )
                             )
                         else:
                             log.log(
@@ -559,9 +561,11 @@ class CopyingManagerWorkerSession(
                                 # have to wait before we send the next request.
                                 pipeline_time = time.time()
                                 self.total_pipelined_requests += 1
-                                self.__pending_add_events_task.next_pipelined_task = self.__get_next_add_events_task(
-                                    copying_params.current_bytes_allowed_to_send,
-                                    for_pipelining=True,
+                                self.__pending_add_events_task.next_pipelined_task = (
+                                    self.__get_next_add_events_task(
+                                        copying_params.current_bytes_allowed_to_send,
+                                        for_pipelining=True,
+                                    )
                                 )
                             else:
                                 pipeline_time = 0.0
@@ -644,8 +648,10 @@ class CopyingManagerWorkerSession(
 
                             # Rate limit based on amount of copied log bytes in a successful request
                             if self.__rate_limiter:
-                                time_slept = self.__rate_limiter.block_until_charge_succeeds(
-                                    log_bytes_sent
+                                time_slept = (
+                                    self.__rate_limiter.block_until_charge_succeeds(
+                                        log_bytes_sent
+                                    )
                                 )
                                 self.__total_rate_limited_time += time_slept
                                 self.__rate_limited_time_since_last_status += time_slept
@@ -800,12 +806,9 @@ class CopyingManagerWorkerSession(
                     > self._last_attempt_time
                     + self.__config.healthy_max_time_since_last_copy_attempt
                 ):
-                    result.health_check_result = (
-                        "Worker session '%s' failed, max time since last copy attempt (%s seconds) exceeded"
-                        % (
-                            self._id,
-                            self.__config.healthy_max_time_since_last_copy_attempt,
-                        )
+                    result.health_check_result = "Worker session '%s' failed, max time since last copy attempt (%s seconds) exceeded" % (
+                        self._id,
+                        self.__config.healthy_max_time_since_last_copy_attempt,
                     )
 
         finally: