+++ tests/unit/builtin_monitors/kubernetes_monitor_test.py
@@ -741,7 +741,9 @@ class TestKubeletApi(BaseScalyrLogCaptureTestCase):
 
         with mock.patch.object(KubeletApi, "_get", fake_get):
             api = KubeletApi(
-                FakeKubernetesApi(), host_ip="127.0.0.1", node_name="FakeNode",
+                FakeKubernetesApi(),
+                host_ip="127.0.0.1",
+                node_name="FakeNode",
             )
 
             self.assertEqual(mock_global_log.error.call_count, 0)
@@ -775,7 +777,9 @@ class TestKubeletApi(BaseScalyrLogCaptureTestCase):
 
         with mock.patch.object(KubeletApi, "_get", fake_get):
             api = KubeletApi(
-                FakeKubernetesApi(), host_ip="127.0.0.1", node_name="FakeNode",
+                FakeKubernetesApi(),
+                host_ip="127.0.0.1",
+                node_name="FakeNode",
             )
 
             self.assertEqual(mock_global_log.warn.call_count, 0)
@@ -894,7 +898,9 @@ class ContainerCheckerTest(TestConfigurationBase):
             ignore_pod_sandboxes=False,
         )
         cc._ContainerChecker__k8s_config_builder = K8sConfigBuilder(
-            config.k8s_log_configs, mock.Mock(), True,
+            config.k8s_log_configs,
+            mock.Mock(),
+            True,
         )
         cc.get_cluster_name = lambda *_: "the-cluster-name"